+++ lib/vialink_flutter_plugin.dart
@@ -2,3 +2,4 @@ library vialink_flutter_plugin;
 
 export 'src/vialink_sdk.dart';
 export 'src/models/deep_link_data.dart';
+export 'src/models/payment_event.dart';